Added users & roles

+++ src/main/java/net/noahk/takenlijst/controllers/ProjectsController.java
@@ -1,10 +1,13 @@
 package net.noahk.takenlijst.controllers;
 
 import net.noahk.takenlijst.dtos.ProjectDto;
+import net.noahk.takenlijst.dtos.ProjectMemberDto;
+import net.noahk.takenlijst.security.MyUserDetails;
 import net.noahk.takenlijst.services.ProjectService;
 import net.noahk.takenlijst.util.Util;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
@@ -28,24 +31,79 @@ public class ProjectsController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ProjectDto> getProject(@PathVariable long id) {
+    public ResponseEntity<Object> getProject(@PathVariable long id, @AuthenticationPrincipal MyUserDetails user) {
         var project = service.getProject(id);
         if (project.isEmpty()) {
             return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
         }
+        if (!service.isProjectMember(user, id)) {
+            return new ResponseEntity<>("You must be a project member to view this project!", HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(project.get());
     }
 
-    @PutMapping("/{id}")
-    public ResponseEntity<String> updateProject(@PathVariable long id, @Valid @RequestBody ProjectDto project, BindingResult bindingResult) {
+    @PostMapping("/assign")
+    public ResponseEntity<String> assignUser(@Valid @RequestBody ProjectMemberDto projectMember, @AuthenticationPrincipal MyUserDetails user, BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
             return Util.getBindingResultResponse(bindingResult);
+        }
+
+        if (service.getProject(projectMember.projectId).isEmpty()) {
+            return new ResponseEntity<>("Project doesn't exist!", HttpStatus.NOT_FOUND);
+        }
+
+        if (!service.isProjectMember(user, projectMember.projectId)) {
+            return new ResponseEntity<>("You must be a project member!", HttpStatus.UNAUTHORIZED);
+        }
+
+        try {
+            service.assignUser(projectMember);
+        } catch (Exception e) {
+            return new ResponseEntity(e.getMessage(), HttpStatus.BAD_REQUEST);
+        }
+        return ResponseEntity.ok("User assigned!");
+    }
+
+    @PostMapping("/unassign")
+    public ResponseEntity<String> unassignUser(@Valid @RequestBody ProjectMemberDto projectMember, @AuthenticationPrincipal MyUserDetails user, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return Util.getBindingResultResponse(bindingResult);
+        }
+
+        if (service.getProject(projectMember.projectId).isEmpty()) {
+            return new ResponseEntity<>("Project doesn't exist!", HttpStatus.NOT_FOUND);
+        }
+
+        if (!service.isProjectMember(user, projectMember.projectId)) {
+            return new ResponseEntity<>("You must be a project member!", HttpStatus.UNAUTHORIZED);
+        }
+
+        try {
+            service.unassignUser(projectMember);
+        } catch (Exception e) {
+            return new ResponseEntity(e.getMessage(), HttpStatus.BAD_REQUEST);
+        }
+        return ResponseEntity.ok("User unassigned!");
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<String> updateProject(@PathVariable long id, @Valid @RequestBody ProjectDto project, @AuthenticationPrincipal MyUserDetails user, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return Util.getBindingResultResponse(bindingResult);
+        }
+
+        if (service.getProject(id).isEmpty()) {
+            return new ResponseEntity<>("Project doesn't exist!", HttpStatus.NOT_FOUND);
+        }
+
+        if (!service.isProjectMember(user, id)) {
+            return new ResponseEntity<>("You must be a project member!", HttpStatus.UNAUTHORIZED);
         }
 
         boolean updated = service.update(id, project);
 
         if (!updated) {
-            return new ResponseEntity<>("Not found!", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>("Something went wrong!", HttpStatus.BAD_REQUEST);
         }
 
         URI uri = URI.create(ServletUriComponentsBuilder.fromCurrentContextPath().path("/projects/" + id).toUriString());
@@ -54,12 +112,23 @@ public class ProjectsController {
     }
 
     @PostMapping("")
-    public ResponseEntity<String> create(@Valid @RequestBody ProjectDto project, BindingResult bindingResult) {
+    public ResponseEntity<String> create(@Valid @RequestBody ProjectDto project, @AuthenticationPrincipal MyUserDetails user, BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
             return Util.getBindingResultResponse(bindingResult);
         }
 
         Long id = service.create(project);
+
+        try {
+            var projectMember = new ProjectMemberDto();
+            projectMember.username = user.getUsername();
+            projectMember.projectId = id;
+
+            service.assignUser(projectMember);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity(e.getMessage(), HttpStatus.BAD_REQUEST);
+        }
 
         URI uri = URI.create(ServletUriComponentsBuilder.fromCurrentContextPath().path("/projects/" + id).toUriString());
 
@@ -67,9 +136,13 @@ public class ProjectsController {
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<?> delete(@PathVariable long id) {
+    public ResponseEntity<?> delete(@PathVariable long id, @AuthenticationPrincipal MyUserDetails user) {
         if (service.getProject(id).isEmpty()) {
-            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>("Project doesn't exist!", HttpStatus.NOT_FOUND);
+        }
+
+        if (!service.isProjectMember(user, id)) {
+            return new ResponseEntity<>("You must be a project member!", HttpStatus.UNAUTHORIZED);
         }
 
         service.delete(id);
